Rename inputs to uniforms
