+++ packages/shader-composer/src/stdlib/index.ts
@@ -4,7 +4,6 @@
 
 export * from "./artistic"
 export * from "./geometry"
-export * from "./input"
 export * from "./logic"
 export * from "./masters"
 export * from "./math"
@@ -12,5 +11,6 @@ export * from "./rotation"
 export * from "./scene"
 export * from "./spaces"
 export * from "./textures"
+export * from "./uniforms"
 export * from "./values"
 export * from "./vectors"
